+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - Add rounding direction to the order book
 - Fix bug in skipping Numia API response errors
 - Ingest tracing
+- Fix multi-hop /custom-direct-quote bug
 - Apply taker fee and spread factor for fill bot amount estimation
 - Reduce fill bot slippage bound by one ulp
 - Fallback to filling each message individually.